Removed Map from encoding to json. Created Service layer and moved the getDateTime function to it
